+++ packages/server/src/modules/middlewares/staticFiles.ts
